@@ -11,72 +11,72 @@
   },
   "private": true,
   "dependencies": {
-    "@alyle/ui": "^2.5.0",
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@beyerleinf/ngx-dnd": "^7.1.1",
-    "@coreui/angular": "^2.4.3",
-    "@coreui/coreui": "^2.1.7",
-    "@coreui/coreui-plugin-chartjs-custom-tooltips": "^1.2.0",
-    "@coreui/icons": "^0.3.0",
-    "@fortawesome/fontawesome-free": "^5.7.2",
-    "@ng-idle/core": "^7.0.0-beta.1",
-    "@ng-idle/keepalive": "^7.0.0-beta.1",
+    "@alyle/ui": "^2.9.6",
+    "@angular/animations": "^8.2.13",
+    "@angular/common": "^8.2.13",
+    "@angular/compiler": "^8.2.13",
+    "@angular/core": "^8.2.13",
+    "@angular/forms": "^8.2.13",
+    "@angular/platform-browser": "^8.2.13",
+    "@angular/platform-browser-dynamic": "^8.2.13",
+    "@angular/router": "^8.2.13",
+    "@beyerleinf/ngx-dnd": "^8.0.3",
+    "@coreui/angular": "~2.6.3",
+    "@coreui/coreui": "^2.1.16",
+    "@coreui/icons": "^0.4.1",
+    "@fortawesome/fontawesome-free": "^5.11.2",
+    "@ng-idle/core": "^8.0.0-beta.4",
+    "@ng-idle/keepalive": "^8.0.0-beta.4",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
-    "ajv": "^6.10.0",
-    "chart.js": "^2.7.3",
-    "chroma-js": "^1.3.6",
-    "core-js": "^2.6.5",
-    "cronstrue": "^1.72.0",
-    "file-saver": "^2.0.1",
+    "ajv": "^6.10.2",
+    "chroma-js": "^2.1.0",
+    "core-js": "^3.3.6",
+    "cronstrue": "^1.84.0",
+    "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.4.1",
     "moment": "^2.24.0",
-    "ng-http-loader": "^5.1.0",
-    "ng-zorro-antd": "^7.5.1",
-    "ngx-bootstrap": "^3.2.0",
-    "ngx-color-picker": "^7.2.15",
-    "ngx-icon-picker": "0.0.9",
-    "ngx-mask": "^7.3.6",
-    "ngx-moment": "^3.2.0",
-    "ngx-perfect-scrollbar": "^7.2.1",
+    "ng-chartjs": "^0.1.2",
+    "ng-http-loader": "^6.0.1",
+    "ng-zorro-antd": "^8.4.1",
+    "ngx-bootstrap": "^5.2.0",
+    "ngx-color-picker": "^8.2.0",
+    "ngx-embed-video": "^1.0.4",
+    "ngx-icon-picker": "0.0.10",
+    "ngx-mask": "^8.1.6",
+    "ngx-moment": "^3.4.0",
+    "ngx-perfect-scrollbar": "^8.0.0",
     "normalize-path": "^3.0.0",
     "pdfobject": "^2.1.1",
-    "popper.js": "^1.14.7",
-    "rxjs": "~6.5.2",
-    "tslib": "^1.9.3",
+    "popper.js": "^1.16.0",
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
     "videogular2": "^7.0.1",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "~7.2.15",
-    "@types/jasmine": "~3.3.9",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "^0.803.17",
+    "@angular/cli": "~8.3.17",
+    "@angular/compiler-cli": "^8.2.13",
+    "@angular/language-service": "~8.2.13",
+    "@types/jasmine": "^3.4.5",
+    "@types/jasminewd2": "^2.0.8",
     "@types/node": "~11.10.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~3.3.0",
+    "codelyzer": "~5.2.0",
+    "insert-line": "^1.1.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~8.0.2",
-    "tslint": "~5.13.1",
-    "typescript": "~3.1.6"
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.4.1",
+    "tslint": "~5.20.0",
+    "typescript": "~3.5.3"
   }
 }
