@@ -20,7 +20,7 @@ import {Subject} from 'rxjs';
 })
 export class ScOptionComponent implements OnChanges {
   changes = new Subject();
-  @ViewChild(TemplateRef) template: TemplateRef<void>;
+  @ViewChild(TemplateRef, {static: false}) template: TemplateRef<void>;
   @Input() scLabel: string;
   // tslint:disable-next-line:no-any
   @Input() scValue: any;
